Fix export with portal with lava slave side
When a portal has a lava texture on the side with the highest segment
number (the slave side) and there are no other lava textures,
the lava texture is ignored but there is a lava collision mesh
created which will be empty. This causes a crash when the empty mesh
is added to the EditorPhysics collision engine.
This commit makes it ignore the empty mesh.

Resolves: #27